#19 Split streaming entities in multiple files and on their own package
