+++ src/main/scala/com/danielasfregola/twitter4s/entities/streaming/LimitTrack.scala
@@ -1,0 +1,5 @@
+package com.danielasfregola.twitter4s.entities.streaming
+
+case class LimitNotice(limit: LimitTrack) extends StreamingMessage
+
+case class LimitTrack(track: Long)
