+++ src/main/scala/com/danielasfregola/twitter4s/entities/streaming/StreamingUpdate.scala
@@ -1,0 +1,5 @@
+package com.danielasfregola.twitter4s.entities.streaming
+
+case class StreamingUpdate(streamingEvent: StreamingMessage)
+
+trait StreamingMessage
