+++ src/main/scala/com/danielasfregola/twitter4s/http/clients/StreamingOAuthClient.scala
@@ -1,7 +1,7 @@
 package com.danielasfregola.twitter4s.http.clients
 
 import akka.actor._
-import akka.pattern.{ ask, pipe }
+import akka.pattern.ask
 import akka.io.IO
 import akka.util.Timeout
 import spray.can.Http
@@ -12,8 +12,8 @@ import spray.httpx.unmarshalling._
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
-
 import com.danielasfregola.twitter4s.entities._
+import com.danielasfregola.twitter4s.entities.streaming._
 
 trait StreamingOAuthClient extends OAuthClient {
