+++ src/main/scala/com/danielasfregola/twitter4s/http/clients/streaming/statuses/parameters/StatusFilterParameters.scala
@@ -5,7 +5,4 @@ import com.danielasfregola.twitter4s.http.marshalling.Parameters
 case class StatusFilterParameters(follow: Option[String] = None,
                                   track: Option[String] = None,
                                   locations: Option[String] = None,
-                                  stall_warnings: Boolean = false) extends Parameters{
-  require(follow.orElse(track).orElse(locations).isDefined,
-    "At least one of 'follow', 'track' or 'locations' needs to be defined")
-}
+                                  stall_warnings: Boolean = false) extends Parameters
